ajout de playlist

+++ public/catalogue/routes/web.php
@@ -27,11 +27,12 @@ Route::delete('deleteMedia/{mediaId}', 'App\Http\Controllers\MediasController@de
 
 Route::get('medias/{id}', 'App\Http\Controllers\MediasController@getMediaDetails');
 
-Route::get('medias/{id}/comments', 'App\Http\Controllers\CommentsController@getMediaComments');
+Route::get('medias/{id}/comments', 'App\Http\Controllers\CommentsController@getMediaComments')->middleware('auth');
 Route::post('medias/{id}/comments', 'App\Http\Controllers\CommentsController@postComment')->middleware('auth');
-Route::delete('medias/{id}/comments', 'App\Http\Controllers\CommentsController@deleteComment');
+Route::delete('medias/{id}/comments', 'App\Http\Controllers\CommentsController@deleteComment')->middleware('auth');
 
-Route::get('myPlaylists/{userId}', 'App\Http\Controllers\PlaylistController@getUserPlaylist');
+Route::get('myPlaylists/{userId}', 'App\Http\Controllers\PlaylistController@getUserPlaylist')->middleware('auth');
+Route::post('myPlaylists/{userId}', 'App\Http\Controllers\PlaylistController@createPlaylist')->middleware('auth')->name('playlist.add');
 
 Route::middleware(['auth:sanctum', 'verified'])->get('/dashboard', function () {
     return view('dashboard');
